skinDetection/requirements.txt
fastapi
Pillow
numpy
tensorflow
gunicorn